// apps/expvi-twitter/src/app/api/model/createPostRequest.ts
/**
 * Expivi Twitter API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreatePostRequest { 
    /**
     * The type of the comment
     */
    type: CreatePostRequest.TypeEnum;
    /**
     * The content of the comment
     */
    text: string;
    /**
     * The url to the video; if post is of type video
     */
    video_url?: string;
    /**
     * The url to the picture; if post is of type picture
     */
    picture_url?: string;
}
export namespace CreatePostRequest {
    export type TypeEnum = 'text' | 'video' | 'picture';
    export const TypeEnum = {
        Text: 'text' as TypeEnum,
        Video: 'video' as TypeEnum,
        Picture: 'picture' as TypeEnum
    };
}